requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.18.0
pydantic>=2.0.0
pyyaml>=6.0

# Data processing and evaluation
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# UI
streamlit>=1.30.0
gradio>=4.0.0  # Optional alternative

# Logging
structlog>=23.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development
black>=23.0.0
flake8>=6.0.0
ipython>=8.12.0
jupyter>=1.0.0
